requirements.txt
python-telegram-bot==21.6
anthropic==0.40.0
openai==1.54.0
google-generativeai==0.8.3
pinecone-client==3.2.2
sqlalchemy==2.0.36
redis==5.2.0
python-dotenv==1.0.1
